test/requirements.txt
pytest==8.3.5
python-dotenv==1.0.1
pytest-order==1.3.0